extract ServiceName
