+++ libs/sabi-core/src/auth/v4/sign/credential_scope.rs
@@ -1,12 +1,13 @@
 use crate::auth::v4::chrono::DateStamp;
-use crate::auth::v4::sign::{ScopeService, ScopeTermination};
+use crate::auth::v4::sign::ScopeTermination;
 use crate::index::AwsRegion;
+use crate::index::ServiceName;
 
 #[derive(Debug)]
 pub struct CredentialScope {
     pub date: DateStamp,
     pub region: AwsRegion,
-    pub service: ScopeService,
+    pub service: ServiceName,
     pub termination: ScopeTermination,
     raw: String,
 }
@@ -15,7 +16,7 @@ impl CredentialScope {
     pub fn from(
         date: DateStamp,
         region: AwsRegion,
-        service: ScopeService,
+        service: ServiceName,
         termination: ScopeTermination,
     ) -> Self {
         let raw = format!(
