+++ libs/sabi-core/src/index/mod.rs
@@ -2,3 +2,4 @@ mod region;
 pub use region::AwsRegion;
 
 mod service;
+pub use service::ServiceName;
