+++ libs/sabi-core/src/index/region.rs
@@ -1,4 +1,3 @@
-
 #[derive(Debug)]
 pub enum AwsRegion {
     ApNorthEast1,
